+++ src/FileSystem/FileSystem.java
@@ -10,13 +10,10 @@ import java.util.Scanner;
 
 public class FileSystem {
 	private static DirectoryStructure root = new DirectoryStructure("root");
-	/** eldeeb part for disk **/
-	private static AllocationStrategy allocationStrategy; /** will be determined in the runtime depends allocation strategy**/
-	private static Scanner read = new Scanner(System.in);/** to read from user **/
-	private static Disk mydisk;
+
 
 	/**
-	 * 
+	 *
 	 * @param path to get the directory in it
 	 * @return directory with this path
 	 */
@@ -24,7 +21,7 @@ public class FileSystem {
 	{
 		if(path.get(0).toLowerCase() != root.getName())
 			return null;
-		
+
 		DirectoryStructure currentDir = root;
 		for(int i=1;i<path.size();i++)
 		{
@@ -32,10 +29,10 @@ public class FileSystem {
 		}
 		return currentDir;
 	}
-	
-	
+
+
 	/**
-	 * 
+	 *
 	 * @param path arraylist of directories' names
 	 * @param name the name of the file
 	 * @param siz the size of the file
@@ -49,17 +46,17 @@ public class FileSystem {
 			System.out.println("Incorrect Path");
 			return false;
 		}
-		
+
 		if(!pathDir.createFile(name, siz))
 			return false;
-		
-		
+
+
 		System.out.println("File created successfully");
 		return true;
 	}
-	
+
 	/**
-	 * 
+	 *
 	 * @param path arraylist of directories' names
 	 * @param name the name of the file
 	 * @return true if deleted and false if not
@@ -72,16 +69,16 @@ public class FileSystem {
 			System.out.println("Incorrect Path");
 			return false;
 		}
-		
+
 		if(!pathDir.deleteFile(name))
 			return false;
-		
+
 		System.out.println("File deleted successfully");
 		return true;
 	}
-	
+
 	/**
-	 * 
+	 *
 	 * @param path arraylist of directories' names
 	 * @param name the name of the file
 	 * @return true if file created successfully and false if not
@@ -94,18 +91,18 @@ public class FileSystem {
 			System.out.println("Incorrect Path");
 			return false;
 		}
-		
+
 		if(!pathDir.createDir(name))
 			return false;
-		
-		
+
+
 		System.out.println("Directory created successfully");
 		return true;
 	}
-	
-	
+
+
 	/**
-	 * 
+	 *
 	 * @param path arraylist of directories' names
 	 * @param name the name of the file
 	 * @return true if deleted and false if not
@@ -118,14 +115,14 @@ public class FileSystem {
 			System.out.println("Incorrect Path");
 			return false;
 		}
-		
+
 		if(!pathDir.deleteDir(name))
 			return false;
-		
+
 		System.out.println("Directory deleted successfully");
 		return true;
 	}
-	
+
 	/**
 	 * Display the disk structure
 	 */
@@ -133,8 +130,8 @@ public class FileSystem {
 	{
 		root.structureDisplay();
 	}
-	
-	
+
+
 	/**
 	 * @return the root directory
 	 */
@@ -142,7 +139,7 @@ public class FileSystem {
 	{
 		return root;
 	}
-	
+
 	/**
 	 * @return JSON String represents the root directory
 	 */
@@ -152,7 +149,7 @@ public class FileSystem {
 		String json = serializer.deepSerialize(this.getRoot());
 		return json;
 	}
-	
+
 	/**
 	 * @param json the JSON string to create the root directory from
 	 */
@@ -160,9 +157,9 @@ public class FileSystem {
 	{
 		this.root = new JSONDeserializer<DirectoryStructure>().deserialize( json );
 	}
-	
+
 	/**
-	 * @param path the path to save the root(JSON) in 
+	 * @param path the path to save the root(JSON) in
 	 * @return true if saved correctly and false if not
 	 */
 	public Boolean save(String path)
@@ -179,7 +176,7 @@ public class FileSystem {
 			return true;
 		return false;
 	}
-	
+
 	/**
 	 * @param path the path to load the root(JSON) from
 	 * @return true if loaded correctly and false if not
@@ -197,83 +194,4 @@ public class FileSystem {
 			return true;
 		return false;
 	}
-	
-	public static void main(String[] args) {
-
-		/** eldeeb part to initialize and prepare the disk and allocation strategy **/
-
-		System.out.print("Enter the size of the disk in KB : ");
-		int diskSize = read.nextInt();
-		System.out.println("Enter the number of allocation choice");
-		System.out.println("(1) Contiguous");
-		System.out.println("(2) Indexed");
-		System.out.println("(3) Liked");
-
-		int choice = read.nextInt();
-		switch (choice){
-			case 1:
-				allocationStrategy = new Contiguous();
-				break;
-			case 2:
-				allocationStrategy = new Indexed();
-				break;
-			case 3:
-				allocationStrategy = new Linked();
-				break;
-		}
-
-		/** mydisk object will deal with disk **/
-		mydisk = Disk.getInstance(allocationStrategy , diskSize);
-
-/**********************************************************************************/
-		/** choice the user can do in our application **/
-		while (true){
-			/** inform the user that he should enter a command **/
-			System.out.print("$$ ");
-			String command = read.nextLine();
-
-			if(command.equals("exit"))
-				break;
-			/**else
-			 	should execute the command ***/
-		}
-/********************************************************************************/
-		FileSystem f = new FileSystem();
-		ArrayList<String> path= new ArrayList<>();
-		path.add("root");
-		f.createFile(path, "File",5);
-		f.createFile(path, "File1",10);
-		f.createFile(path, "File2",12);
-		f.createDir(path, "New Folder");
-		
-		path.add("New Folder");
-		f.createDir(path, "Sherif");
-		f.createFile(path, "LOL", 5);
-
-		path.add("sherif");
-		f.createFile(path, "LOL", 5);
-		f.createFile(path, "Good", 5);
-		f.createDir(path, "Good");
-
-		f.displayDiskStructure();
-		
-		
-		f.deleteFile(path, "lol");
-		f.displayDiskStructure();
-
-		// remove last element
-		path.remove(path.size()-1);
-		f.deleteDir(path, "sherif");
-		f.displayDiskStructure();
-
-		String filePath = "files/fileSystem.json";
-		f.save(filePath);
-		f.load(filePath);
-		System.out.println();
-		f.displayDiskStructure();
-
-
-	}
-	
-
 }
