--- src/Parser/CreateFolderAndDeleteCommand.java
@@ -1,13 +1,0 @@
-package Parser;
-
-import java.util.ArrayList;
-
-public class CreateFolderAndDeleteCommand extends CMD {
-    public String fileName ;
-    public ArrayList<String> path ;
-
-    public CreateFolderAndDeleteCommand()
-    {
-        path = new ArrayList<>() ;
-    }
-}
